src/requirements.txt
pgvector==0.1.7
tiktoken==0.4.0
langchain==0.0.166
python-dotenv==1.0.0
openai==0.27.6
psycopg2-binary==2.9.6
streamlit==1.22.0
